change the Network class
